Add Bob Jenkin's spooky hash
Spooky is an excellent hash on 64bit machines, optimized
for large keys. This adds a slightly modified variant
of the spooky V1 reference code, with some minor changes
from V2. Source is
http://burtleburtle.net/bob/hash/spooky.html

I modified it to conform (mostly) to GNU coding standards
and added some compat interfaces to make it look similar
to the incremential murmurhash.

The original code is in the Public Domain, I don't own it.

XXX improvements to merge()