Add a spooky implementation for inchash for 64bit hosts
Reimplement the inchash interface with the spooky hash for 64bit
hosts.   Spooky is optimized for long hashes and avoids redoing a
full hash computation for every input value. It's also optimized for 64bit.
This allows us to exploit the full 64bit computing capability.
For 32bit we stay with the old jhash.

XXX numbers

gcc/:

2014-07-27  Andi Kleen  <ak@linux.intel.com>

	* inchash.h (inchash): Rename to inchash_jhash.
	(inchash_spoody): Add a new spooky hash based implementation.
	(inchash): Add new child class with common functions.
	(add_object): Move to template function in inchash.