Support slim LTO bootstrap
Add a new bootstrap-lto-slim config file that enables slim (non-fat) LTO
bootstrap. This improves the performance of the LTO bootstrap.

This requires using gcc-ar/ranlib in post stage 1 builds, so these
are passed to all sub builds.

I made it a new file for now because it requires the host to support
linker plugins, which I believe is not supported everywhere yet?

XXX numbers

config/:

2014-08-04  Andi Kleen  <ak@linux.intel.com>

	* bootstrap-lto-slim.mk: New file.

/:

2014-08-04  Andi Kleen  <ak@linux.intel.com>

	* Makefile.tpl (POSTSTAGE1_HOST_EXPORTS): Add LTO_EXPORTS.
	POSTSTAGE1_FLAGS_TO_PASS):  Add LTO_FLAGS_TO_PASS.
	* Makefile.in: Regenerate.

gcc/:

2014-08-04  Andi Kleen  <ak@linux.intel.com>

	* doc/install.texi: Documentation bootstrap-lto-slim