Add a pass to automatically add ptwrite instrumentation
[v3: addressing review comments and fixing more bugs.
Instrumentation is now based on generic walkers and SSA,
which minimizes redundancies and avoids corner cases.
Now able to trace floats/doubles too. Some fixes for issues from
running the full test suite with -fvartrace=all, mainly
by not instrumenting special cases like throwing memory
accesses or asm goto. Improved documentation.]

Add a new pass to automatically instrument changes to variables
with the new PTWRITE instruction on x86. PTWRITE writes a 4 or 8 byte
field into an Processor Trace log, which allows low over head
logging of information. Essentially it's a hardware accelerated
printf.

This allows to reconstruct how values later from the log,
which can be useful for debugging or other analysis of the program
behavior. With the compiler support this can be done with without
having to manually add instrumentation to the code.

Using dwarf information this can be later mapped back to the variables.
The decoder decodes the PTWRITE instructions using IP information
in the log, and then looks up the argument in the debug information.
Then this can be used to reconstruct the original variable
name to display a value history for the variable.

There are new options to enable instrumentation for different types,
and also a new attribute to control analysis fine grained per
function or variable level. The attributes can be set on both
the variable and the type level, and also on structure fields.
This allows to enable tracing only for specific code in large
programs in a flexible matter.

The pass is generic, but only the x86 backend enables the necessary
hooks. When the backend enables the necessary hooks (with -mptwrite)
there is an additional pass that looks through the code for
attribute vartrace enabled functions or variables.

Currently the code can be tested with SDE, or on a Intel
Gemini Lake system with a new enough Linux kernel (v4.10+)
that supports PTWRITE for PT. Gemini Lake is used in low
end laptops ("Intel Pentium Silver J5...... / Celeron N4... /
Celeron J4...")

Linux perf (4.10+) can be used to record the values

perf record -e intel_pt/ptw=1,pt=1,branch=0,fup_on_ptw=1/u ./program
perf script -F +srcline ..

I have an experimential version of perf that can also use
dwarf information to symbolize many[1] values back to their variable
names. So far it is not in standard perf, but available at

https://git.kernel.org/pub/scm/linux/kernel/git/ak/linux-misc.git/log/?h=perf/var-resolve-4

It is currently not able to decode all variable locations to names,
but a large subset.

The CPU can potentially generate very data high bandwidths when
code doing a lot of computation is heavily instrumented.
This can cause some data loss in both the CPU and also in perf
logging the data when the disk cannot keep up.

Running some larger workloads most workloads do not cause
CPU level overflows, but I've seen it with -fvartrace
with crafty, and with more workloads with -fvartrace=locals.

Recommendation is to not fully instrument programs,
but only areas of interest either at the file level or using
the attributes.

perf and the disk often cannot keep up
with the data bandwidth for longer computations. In this case
it's possible to use perf snapshot mode (add --snapshot
to the command line above). The data will be only logged to
a memory ring buffer then, and only dump the buffers on events
of interest by sending SIGUSR2 to the perf binrary.

In the future this will be hopefully better supported with
core files and gdb.

Passes bootstrap and test suite on x86_64-linux, also
bootstrapped and tested gcc itself with full -fvartrace
and -fvartrace=all instrumentation, and running
the test suite with -fvartrace=all. There are some
additional failures in the test suite with -fvartrace=all, but they
all seem to be due to fragile tests not expecting additional
code.

gcc/:

2018-11-24  Andi Kleen  <ak@linux.intel.com>
	    Richard Biener  <rguenther@suse.de>

	* Makefile.in: Add tree-vartrace.o.
	* common.opt: Add -fvartrace
	* opts.c (parse_vartrace_options): Add.
	(common_handle_option): Call parse_vartrace_options.
	* config/i386/i386.c (ix86_vartrace_func): Add.
	(TARGET_VARTRACE_FUNC): Add.
	* doc/extend.texi: Document vartrace/no_vartrace
	attributes.
	* doc/invoke.texi: Document -fvartrace.
	* doc/tm.texi (TARGET_VARTRACE_FUNC): Add.
	* passes.def: Add vartrace pass.
	* target.def (vartrace_func): Add.
	* tree-pass.h (make_pass_vartrace): Add.
	* tree-vartrace.c: New file to implement vartrace pass.

gcc/c-family/:

2018-11-24  Andi Kleen  <ak@linux.intel.com>

	* c-attribs.c (handle_vartrace_attribute,
	  handle_no_vartrace_attribute): New functions.
	  (attr_vartrace_exclusions): Add.

config/:

2018-11-24  Andi Kleen  <ak@linux.intel.com>

	* bootstrap-vartrace.mk: New.
	* bootstrap-vartrace-locals.mk: New.