Add a pass to automatically add ptwrite instrumentation
XXX add urls, sde

Add a new pass to automatically instrument changes to variables
with the new PTWRITE instruction on x86. PTWRITE writes a 4 or 8 byte
field into an external Processor Trace log.

This allows to reconstruct how values later, which can be useful for
debugging or other analysis of the program behavior. With the compiler
support this can be done with without having to manually add instrumentation
to the code.

Using dwarf information this can be later mapped back to the variables.

There are new options to enable instrumentation for different types,
and also a new attribute to control analysis fine grained per
function or variable level. The attributes can be set on both
the variable and the type level, and also on structure fields.
This allows to enable tracing only for specific code in large
programs.

The pass is generic, but only the x86 backend enables the necessary
hooks. When the backend enables the necessary hooks (with -mptwrite)
there is an additional pass that looks through the code for
attribute vartrace enabled functions or variables.

The -fvartrace-locals options is experimental: it works, but it
generates redundant ptwrites. This could be optimized later.

Currently the code can be tested with SDE:

I put a decoding script at ...
It requires a patched perf patch
This perf version is also able to reconstruct the many[1] values
from a real PT log and symbolize them using dwarf information.

[1] Many: so far it only supports register variables that are not
arguments.

gcc/:

2017-11-11  Andi Kleen  <ak@linux.intel.com>

	* Makefile.in: Add tree-vartrace.o.
	* common.opt: Add -fvartrace, -fvartrace-returns,
	-fvartrace-args, -fvartrace-reads, -fvartrace-writes,
	-fvartrace-locals
	* config/i386/i386.c (ix86_vartrace_func): Add.
	(TARGET_VARTRACE_FUNC): Add.
	* doc/extend.texi: Document vartrace/no_vartrace
	attributes.
	* doc/invoke.texi: Document -fvartrace, -fvartrace-returns,
	-fvartrace-args, -fvartrace-reads, -fvartrace-writes,
	-fvartrace-locals
	* doc/tm.texi (TARGET_VARTRACE_FUNC): Add.
	* passes.def: Add vartrace pass.
	* target.def (vartrace_func): Add.
	* tree-pass.h (make_pass_vartrace): Add.
	* tree-vartrace.c: New file to implement vartrace pass.

gcc/c-family/:

2017-11-11  Andi Kleen  <ak@linux.intel.com>

	* c-attribs.c (handle_vartrace_attribute): New function.