Add a pass to automatically add ptwrite instrumentation
[v4: Rebased on current tree. Avoid some redundant log statements
for locals and a few other fixes.  Fix some comments. Improve
documentation. Did some studies on the debug information quality,
see below]

Add a new pass to automatically instrument changes to variables
with the new PTWRITE instruction on x86. PTWRITE writes a 4 or 8 byte
field into an Processor Trace log, which allows low over head
logging of information. Essentially it's a hardware accelerated
printf.

This allows to reconstruct how values later from the log,
which can be useful for debugging or other analysis of the program
behavior. With the compiler support this can be done with without
having to manually add instrumentation to the code.

Using dwarf information this can be later mapped back to the variables.
The decoder decodes the PTWRITE instructions using IP information
in the log, and then looks up the argument in the debug information.
Then this can be used to reconstruct the original variable
name to display a value history for the variable.

There are new options to enable instrumentation for different types,
and also a new attribute to control analysis fine grained per
function or variable level. The attributes can be set on both
the variable and the type level, and also on structure fields.
This allows to enable tracing only for specific code in large
programs in a flexible matter.

The pass is generic, but only the x86 backend enables the necessary
hooks. When the backend enables the necessary hooks (with -mptwrite)
there is an additional pass that looks through the code for
attribute vartrace enabled functions or variables.

Earlier there were concerns that the debug information is not
always associated with the ptwrite instruction because the
backend doesn't know how to keep it together.

I did some experiments using -fdump-rtl-final, just checking if the
PTWRITE builtin has a variable location on "loop-unroll.c" from
the gcc source.

With -fvartrace=args there is good coverage, the ptwrite always had a
usable variable name.

With -fvartrace=returns there is usually no variable name, but in this
case the decoder can figure it out by looking for the RET, and knowing
that the value is in %rax.

With -fvartrace=reads,writes the ptwrite usually just has the variable
name of the gimple temporary in a register.
However there is near always a rtl set for the address just before it,
and the set tends to have the expected name/type and offset (if for a
structure). There are two ways to handle this: either could teach the
decoder to track debug info for registers. Or alternatively could
change the ptwrite define_insn to avoid splitting the effective
address into a separate register. I tried this with some different
constraints, but wasn't successfull. I hope there's some way to do
this though.

With -fvartrace=locals there's a mix. Most accesses have the correct
variable name, but sometimes it is lost. I believe that's acceptable,
locals is more an experimental option and may not be too useful anyways
because it generates a lot of traffic.

Currently the code can be tested with SDE, or on a Intel
Gemini Lake system with a new enough Linux kernel (v4.10+)
that supports PTWRITE for PT. Gemini Lake is used in low
end laptops ("Intel Pentium Silver J5...... / Celeron N4... /
Celeron J4...")

Linux perf (4.10+) can be used to record the values

perf record -e intel_pt/ptw=1,pt=1,branch=0,fup_on_ptw=1/u ./program
perf script -F +srcline ..

I have an experimential version of perf that can also use
dwarf information to symbolize many[1] values back to their variable
names. So far it is not in standard perf, but available at

https://git.kernel.org/pub/scm/linux/kernel/git/ak/linux-misc.git/log/?h=perf/var-resolve-5

It is currently not able to decode all variable locations to names,
but a large subset.

The CPU can potentially generate very data high bandwidths when
code doing a lot of computation is heavily instrumented.
This can cause some data loss in both the CPU and also in perf
logging the data when the disk cannot keep up.

Running some larger workloads most workloads do not cause
CPU level overflows, but I've seen it with -fvartrace
with crafty, and with more workloads with -fvartrace=locals.

Recommendation is to not fully instrument programs,
but only areas of interest either at the file level or using
the attributes.

perf and the disk often cannot keep up
with the data bandwidth for longer computations. In this case
it's possible to use perf snapshot mode (add --snapshot
to the command line above). The data will be only logged to
a memory ring buffer then, and only dump the buffers on events
of interest by sending SIGUSR2 to the perf binrary.

In the future this will be hopefully better supported with
core files and gdb.

Passes bootstrap and test suite on x86_64-linux, also
bootstrapped and tested gcc itself with full -fvartrace
and -fvartrace=all instrumentation, and running
the test suite with -fvartrace=all. There are some
additional failures in the test suite with -fvartrace=all, but they
all seem to be due to fragile tests not expecting additional
code.

gcc/:

2019-11-10  Andi Kleen  <ak@linux.intel.com>
	    Richard Biener  <rguenther@suse.de>

	* Makefile.in: Add tree-vartrace.o.
	* common.opt: Add -fvartrace
	* opts.c (parse_vartrace_options): Add.
	(common_handle_option): Call parse_vartrace_options.
	* config/i386/i386.c (ix86_vartrace_func): Add.
	(TARGET_VARTRACE_FUNC): Add.
	* doc/extend.texi: Document vartrace/no_vartrace
	attributes.
	* doc/invoke.texi: Document -fvartrace.
	* doc/tm.texi (TARGET_VARTRACE_FUNC): Add.
	* passes.def: Add vartrace pass.
	* target.def (vartrace_func): Add.
	* tree-pass.h (make_pass_vartrace): Add.
	* tree-vartrace.c: New file to implement vartrace pass.

gcc/c-family/:

2019-11-10  Andi Kleen  <ak@linux.intel.com>

	* c-attribs.c (handle_vartrace_attribute,
	  handle_no_vartrace_attribute): New functions.
	  (attr_vartrace_exclusions): Add.

config/:

2019-11-10  Andi Kleen  <ak@linux.intel.com>

	* bootstrap-vartrace.mk: New.
	* bootstrap-vartrace-locals.mk: New.